refactor(text) improve reusability
