+++ packages/core/src/components/text/index.ts
@@ -19,7 +19,8 @@ export * from './Headline6';
 export * from './Overline';
 export * from './Paragraph1';
 export * from './Paragraph2';
-export * from './RfxText';
+export * from './renderRfxTextComponent';
+export * from './renderTextComponent';
 export * from './RfxTextProps';
 export * from './RfxTextTheme';
 export * from './RfxTextVariantsTheme';
