+++ packages/core/src/components/text/renderTextComponent.tsx
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) Flavio Silva https://flsilva.com
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import * as React from 'react';
+import { Text, TextProps } from 'react-native';
+
+import { extractTextProps } from '../../utils/props';
+import { BuiltInSimpleComponentProps } from '../BuiltInSimpleComponentProps';
+
+export const renderTextComponent = <
+  ComponentProps extends TextProps & Readonly<{ children?: React.ReactNode }>
+>(
+  props: ComponentProps,
+  Component:
+    | typeof Text
+    | React.ComponentType<BuiltInSimpleComponentProps<ComponentProps>> &
+        TextProps = Text,
+): JSX.Element => {
+  const textProps = extractTextProps(props);
+
+  if (Component === Text) {
+    return <Component {...textProps}>{props.children}</Component>;
+  }
+
+  return (
+    <Component complexComponentProps={props} {...textProps}>
+      {props.children}
+    </Component>
+  );
+};
